Increase timeout
When the date-picker is not visible (e.g. because it's in a pop-up) the timeout-loop will take up all available processing time making CPU fans spin up for naught. Checking ten times per second instead of as fast as possible is sufficient - probably still too much but bearable, at least.